Converts ssh ciphers and kex vars to lists
The vars represent information that is best understood as a list, and
therefore should be stored as a list, and interpolated as a
comma-separated string only at the presentation layer, when rendering
the template.

Storing the vars as lists is eminently more readable, which encourages
a deeper understanding of this role, and of the SSH service in general.