`ssh_print_motd` replaced by `ssh_print_pam_motd`
Signed-off-by: jeanmonet <jeanmonet@gmail.com>

This may be a leftover in README.md from a previous change of the ssh_hardening role.
I assume that the line `ssh_print_motd` needs to be deleted from the readme. I think (but have not checked) this option was replaced by `ssh_print_pam_motd`. I assume this because it appears that's the only 'motd' variable used in the hardening task:
https://github.com/dev-sec/ansible-collection-hardening/blob/83e29b01f529be07a18f5223427258e986cb6e2b/roles/ssh_hardening/tasks/hardening.yml#L57